requirements.txt
fastapi
uvicorn[standard]
gunicorn
sqlmodel
passlib[bcrypt]
python-jose[cryptography]
